src/utils/config.py
"""Configuration management utilities."""
import yaml
from pathlib import Path
from typing import Dict, Any


def load_config(config_path: str) -> Dict[str, Any]:
    """
    Load YAML configuration file.

    Args:
        config_path: Path to YAML config file

    Returns:
        Configuration dictionary

    TODO: Implement config loading with validation
    """
    pass  # To be implemented


def merge_configs(*configs: Dict) -> Dict[str, Any]:
    """
    Merge multiple configuration dictionaries.

    Args:
        *configs: Variable number of config dicts

    Returns:
        Merged configuration

    TODO: Implement config merging
    """
    pass  # To be implemented
