requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.13.0
tensorflow-hub>=0.14.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
audiomentations>=0.31.0
pydub>=0.25.0

# Biometric Processing
neurokit2>=0.2.4
hrv-analysis>=1.0.4
scipy>=1.11.0
numpy>=1.24.0
pandas>=2.0.0

# Deep Learning
transformers>=4.30.0
scikit-learn>=1.3.0

# API & Backend
fastapi>=0.103.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Explainability
shap>=0.42.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Development
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
tqdm>=4.66.0
matplotlib>=3.7.0
seaborn>=0.12.0
